requirements.txt
# Core dependencies
ultralytics>=8.0.0       # For YOLOv8 models
torch>=1.12.0            # PyTorch for deep learning
torchvision>=0.13.0      # Vision utilities for PyTorch

# Data handling and processing
numpy>=1.21.0            # Numerical computing
pandas>=1.3.0            # Data analysis tools
PyYAML>=6.0              # YAML file parsing

# Visualization
matplotlib>=3.4.0        # Plotting and visualization
seaborn>=0.11.2          # Enhanced visualization

# Utilities
tqdm>=4.64.0             # Progress bars
Pillow>=9.0.0            # Image processing
opencv-python>=4.6.0     # Computer vision utilities

# Optional - for efficiency
# psutil>=5.9.0          # System monitoring
# py-cpuinfo>=8.0.0      # CPU information
